yolo_cam/requirements.txt
sqlalchemy
jinja2
python-multipart
opencv-python
face-recognition
ultralytics
python-dotenv
cryptography